feat(34-02): split blog and blog category pages with generateMetadata and pagination noindex
